Set up the project with Webpack
