+++ .gitignore
@@ -1,1 +1,5 @@
 web.config
+node_modules/
+dist/bundle.js
+dist/bundle.js.map
+*.log
