--- index.html
@@ -1,56 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-      <meta charset="UTF-8">
-      <meta http-equiv="X-UA-Compatible" content="IE=edge">
-      <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-      <title>Oh Hell!</title>
-
-      <!-- angular -->
-      <script src="https://cdnjs.cloudflare.com/ajax/libs/angular.js/1.6.7/angular.min.js" integrity="sha256-bs1GvtqRJyWn3a6f6BiFOdX9LK86qR4UuPKmLr5JXTE=" crossorigin="anonymous"></script>
-      <script src="https://cdnjs.cloudflare.com/ajax/libs/angular.js/1.6.7/angular-route.min.js" integrity="sha256-8bgkuoYyDpASEZz5bIftIpnq9jPDnui2P34vD6i9Ae4=" crossorigin="anonymous"></script>
-
-      <!-- smart table -->
-      <script src="https://cdnjs.cloudflare.com/ajax/libs/angular-smart-table/2.1.11/smart-table.min.js" integrity="sha256-+Vv2BYQhkWVvDxvLEuCHqx/aKOsS/CEIfBRG+w3aMzs=" crossorigin="anonymous"></script>
-
-      <!-- bootstrap -->
-      <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.0.0-beta.2/css/bootstrap-reboot.min.css" integrity="sha256-qjUPFjgnGztVhVUkJRywzCtVv7KW9GM0bQcEZglyS/w=" crossorigin="anonymous" />
-      <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta.3/css/bootstrap.min.css" integrity="sha384-Zug+QiDoJOrZ5t4lssLdxGhVrurbmBWopoEl+M6BdEfwnCJZtKxi1KgxUyJq13dy" crossorigin="anonymous">
-
-      <!-- chart.js -->
-      <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.7.1/Chart.bundle.min.js" integrity="sha256-N4u5BjTLNwmGul6RgLoESPNqDFVUibVuOYhP4gJgrew=" crossorigin="anonymous"></script>
-
-      <!-- lodash -->
-      <script src="https://cdnjs.cloudflare.com/ajax/libs/lodash.js/4.17.4/lodash.min.js" integrity="sha256-8E6QUcFg1KTnpEU8TFGhpTGHw5fJqB9vCms3OhAYLqw=" crossorigin="anonymous"></script>
-
-      <!-- localforage -->
-      <script src="https://cdnjs.cloudflare.com/ajax/libs/localforage/1.5.5/localforage.min.js" integrity="sha256-AQ34lTQ8NC/rakAt7/GcayJkOJztnTsyQL4Kz7J86TY=" crossorigin="anonymous"></script>
-
-      <!-- moment -->
-      <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.20.1/moment.min.js" integrity="sha256-ABVkpwb9K9PxubvRrHMkk6wmWcIHUE9eBxNZLXYQ84k=" crossorigin="anonymous"></script>
-
-      <!-- local -->
-      <link rel="stylesheet" href="styles/index.css" />
-      <link rel="stylesheet" href="styles/icons.css" />
-
-      <script src="index.js" type="module"></script>
-  </head>
-
-  <body ng-controller="AppController as App" ng-class="{'dark': App.darkTheme}">
-      <div class="topNav">
-          <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
-              <a class="navbar-brand" href="#">Bloody Hell!</a>
-              <form class="form-inline ml-auto">
-                  <a class="btn btn-primary btn-sm" href="#!/history">Game History</a>
-                  <button class="btn btn-primary btn-sm" ng-click="App.toggleScoreView()" ng-disabled="App.game.isFinished">Toggle Score View</button>
-                  <button class="btn btn-primary btn-sm" ng-click="App.toggleTheme()">Toggle Theme</button>
-                  <a class="btn btn-primary btn-sm" href="#!/game/0">Create New Game</a>
-              </form>
-          </nav>
-      </div>
-      <div class="container-fluid">
-          <div ng-view></div>
-      </div>
-  </body>
-</html>
