+++ src/game/game-controller.js
@@ -1,3 +1,5 @@
+import _ from 'lodash';
+
 import { GameService } from '../services/game.js';
 import { ChartService } from '../services/chart.js';
 import { StorageService } from '../services/storage.js';
