+++ src/history/history-controller.js
@@ -1,3 +1,6 @@
+import _ from 'lodash';
+import moment from 'moment';
+
 import { GameService } from '../services/game.js';
 import { ChartService } from '../services/chart.js';
 import { StorageService } from '../services/storage.js';
