+++ src/routes.js
@@ -4,12 +4,12 @@ import { HistoryController } from './history/history-controller.js';
 export function routes($routeProvider) {
   $routeProvider
     .when('/game/:id?', {
-      templateUrl: './app/game/game-template.html',
+      template: require('./game/game-template.html'),
       controller: GameController,
       controllerAs: 'Game'
     })
     .when('/history', {
-      templateUrl: './app/history/history-template.html',
+      template: require('./history/history-template.html'),
       controller: HistoryController,
       controllerAs: 'History'
     })
