+++ webpack.config.js
@@ -1,0 +1,42 @@
+const path = require('path');
+
+module.exports = {
+  entry: './src/index.js',
+  output: {
+    filename: "bundle.js",
+    path: path.resolve(__dirname, 'dist')
+  },
+  devServer: {
+    port: 8081
+  },
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        exclude: /node_modules/,
+        use: {
+          loader: 'babel-loader',
+            options: {
+              presets: ['env']
+          }
+        }
+      },
+      {
+        test: /\.css$/,
+        use: ['style-loader', 'css-loader']
+      },
+      {
+        test: /\.(ttf|eot|woff|woff2|svg)$/,
+        loader: "file-loader",
+        options: {
+          name: "fonts/[name].[ext]",
+        },
+      },
+      {
+        test: /\.html$/,
+        loader: 'html-loader'
+      }
+    ]
+  },
+  devtool: 'source-map'
+}
